Update 6.NoninvasivePrenatalDiagnosis.md
Additional resources are needed for these jobs. It looks like this is the first time in the semester that the defaults are not sufficient. May be worth mentioning in class.

The calculations at the end were behaving oddly trying to cat | grep | awk so i just removed the cat step. This seems to work just fine.